requirements.txt
appdirs==1.4.4
asgiref==3.6.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
Brotli==1.0.9
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
command-not-found==0.3
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
filelock==3.9.0
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
ipython==7.31.1
jedi==0.18.0
jeepney==0.7.1
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
lz4==3.1.3+dfsg
matplotlib==3.5.1
matplotlib-inline==0.1.3
more-itertools==8.10.0
mpmath==0.0.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
outcome==1.2.0
packaging==21.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.6.2
ply==3.11
prompt-toolkit==3.0.28
ptyprocess==0.7.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PySocks==1.7.1
python-apt==2.4.0
python-dateutil==2.8.1
python-dotenv==0.21.0
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
requests==2.28.2
scipy==1.8.0
SecretStorage==3.3.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.3
sympy==1.9
systemd-python==234
tqdm==4.64.1
traitlets==5.1.1
trio==0.22.0
trio-websocket==0.9.2
ubuntu-advantage-tools==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.14
virtualenv==20.17.1
wadllib==1.3.6
wcwidth==0.2.5
webdriver-manager==3.8.5
webencodings==0.5.1
wsproto==1.2.0
zipp==1.0.0
